fix: mu50のXG Parameter Changeを既存のフォルダに統合
