+++ main.ts
@@ -113,37 +113,6 @@ const drumSetList = new Domino.DrumSetList([
   elxxxDrumMap,
 ]);
 
-// mu50.xmlのControlChangeMacroListの一部を統合
-const f = await Deno.readFile("./memo/mu50.xml");
-const mu50Str = Encoding.convert(f, {
-  to: "UTF8",
-  from: "SJIS",
-  type: "string",
-});
-const mu50 = Domino.File.fromXML(mu50Str);
-const mu50CcmList = mu50.moduleData.controlChangeMacroList;
-if (!mu50CcmList) {
-  throw new Error("mu50.xml ControlChangeMacroList is not found");
-}
-function filterCCM(ccm: Domino.CCMFolder | Domino.ControlChangeMacroList) {
-  return ccm.tags.filter((tag) => {
-    if (
-      tag instanceof Domino.CCM &&
-      (tag.param.id < 140 || tag.param.id === 200 || tag.param.id === 210 ||
-        tag.param.id === 275)
-    ) {
-      return false;
-    }
-    if (tag instanceof Domino.CCMFolder) {
-      tag.tags = filterCCM(tag);
-      if (tag.tags.length === 0) return false;
-    }
-    return true;
-  });
-}
-const addCcmList = filterCCM(mu50CcmList);
-ccmList.tags.push(...addCcmList);
-
 const file = new Domino.File({
   name: "Electone",
   folder: "YAMAHA",
